Lista DAO implementation

+++ src/main/java/edu/curso/goodooit/domain/repository/IComentarioDAO.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface IComentarioDAO {
     public List<Comentario> buscarTodosComentarios() throws SQLException;
 
-    public Comentario buscarComentarioIdUsuario(Long idUsuario) throws SQLException;
+    public Comentario buscarComentarioIdUsuario(Integer idUsuario) throws SQLException;
 
-    public Comentario buscarComentarioIdTarefa(Long idTarefa) throws SQLException;
+    public Comentario buscarComentarioIdTarefa(Integer idTarefa) throws SQLException;
 
     public Comentario registrarComentario(Comentario comentario) throws SQLException;
